# perf(portex): cache "Signature" for "PortexExternalType"