| cases                  | before              | after              |
| ---------------------- | ------------------- | ------------------ |
| ob.label.Box2D()       | 153 µs ± 7.57 µs    | 122 µs ± 1.41 µs   |
| ob.label.Box3D()       | 273 µs ± 3.3 µs     | 246 µs ± 3.32 µs   |