@GravitiRob 2022-05-26
Task linked: [CU-2dnbxwy 优化 PortexType.__init__ 方法，提高初始化效率](https://app.clickup.com/t/2dnbxwy)